User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -61,6 +61,12 @@ class ListingController extends Controller
 
     // update listing
     public function update(Request $request, Listing $listing) {
+
+        // Check if logged in user is the owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -82,6 +88,12 @@ class ListingController extends Controller
 
     // Delete listing
     public function destroy(Listing $listing) {
+
+        // Check if logged in user is the owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
